+++ Java/core/src/main/java/com/amazon/randomcutforest/state/tree/CompactRandomCutTreeState.java
@@ -25,7 +25,7 @@ import com.amazon.randomcutforest.tree.CompactRandomCutTreeDouble;
 public class CompactRandomCutTreeState {
     public LeafStoreState leafStoreState;
     public NodeStoreState nodeStoreState;
-    public short rootIndex;
+    public int rootIndex;
 
     public CompactRandomCutTreeState() {
